@@ -3,24 +3,25 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "axios": "^0.21.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "@testing-library/jest-dom": "^6.4.5",
+    "@testing-library/react": "^15.0.7",
+    "@testing-library/user-event": "^14.5.2",
+    "axios": "^1.6.8",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-icon": "^1.0.0",
     "react-icons": "^4.2.0",
     "react-infinite-scroll-component": "^6.1.0",
-    "react-redux": "^7.2.4",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "react-select": "^4.3.1",
-    "redux": "^4.1.1",
+    "react-redux": "^9.1.2",
+    "react-router-dom": "^6.23.0",
+    "react-scripts": "^5.0.1",
+    "react-select": "^5.8.0",
+    "redux": "^5.0.1",
     "redux-infinite-scroll": "^1.0.9",
     "redux-persist": "^6.0.0",
-    "redux-thunk": "^2.3.0",
-    "styled-components": "^5.3.0",
-    "web-vitals": "^1.0.1"
+    "redux-thunk": "^3.1.0",
+    "styled-components": "^6.1.9",
+    "web-vitals": "^3.5.2"
   },
   "scripts": {
     "start": "react-scripts start",
