@@ -1,5 +1,5 @@
 import { Provider } from "react-redux";
-import {BrowserRouter as Router, Route} from "react-router-dom";
+import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 import Characters from "./components/pages/characters";
 import Store from "./store";
 import Favourites from "./components/pages/favourites";
@@ -7,20 +7,21 @@ import Character from "./components/pages/character";
 import SearchResults from "./components/pages/searchResults";
 
 const App = () => {
-
   return (
     <div className="App">
       <Router>
         <Provider store={Store}>
-          <Route path="/characters/:id" component={Character}/>
-          <Route path="/" exact component={Characters}/>
-          <Route path="/characters" exact component={Characters}/>
-          <Route path="/favourites" component={Favourites}/>
-          <Route path="/search/:keyword" component={SearchResults} />
+          <Routes>
+            <Route path="/" exact element={<Characters />} />
+            <Route path="/characters/:id" element={<Character />} />
+            <Route path="/characters" exact element={<Characters />} />
+            <Route path="/favourites" element={<Favourites />} />
+            <Route path="/search/:keyword" element={<SearchResults />} />
+          </Routes>
         </Provider>
       </Router>
     </div>
   );
-}
+};
 
 export default App;
